+++ tics/geomath/src/vector3d.cpp
@@ -6,7 +6,7 @@ Vector3D Vector3D::operator+(const Vector3D & rhs) const {
 	return { x + rhs.x, y + rhs.y, z + rhs.z };
 }
 
-Vector3D Vector3D::operator+(const float & rhs) const {
+Vector3D Vector3D::operator+(const double & rhs) const {
 	return { x + rhs, y + rhs, z + rhs };
 }
 
@@ -21,7 +21,7 @@ Vector3D Vector3D::operator-(const Vector3D & rhs) const {
 	return { x - rhs.x, y - rhs.y, z - rhs.z };
 }
 
-Vector3D Vector3D::operator-(const float & rhs) const {
+Vector3D Vector3D::operator-(const double & rhs) const {
 	return { x - rhs, y - rhs, z - rhs };
 }
 
@@ -36,7 +36,7 @@ Vector3D Vector3D::operator*(const Vector3D & rhs) const {
 	return { x * rhs.x, y * rhs.y, z * rhs.z };
 }
 
-Vector3D Vector3D::operator*(const float & rhs) const {
+Vector3D Vector3D::operator*(const double & rhs) const {
 	return { x * rhs, y * rhs, z * rhs };
 }
 
@@ -51,7 +51,7 @@ Vector3D Vector3D::operator/(const Vector3D & rhs) const {
 	return { x / rhs.x, y / rhs.y, z / rhs.z };
 }
 
-Vector3D Vector3D::operator/(const float & rhs) const {
+Vector3D Vector3D::operator/(const double & rhs) const {
 	return { x / rhs, y / rhs, z / rhs };
 }
 
@@ -62,6 +62,6 @@ Vector3D & Vector3D::operator/=(const Vector3D & rhs) {
 	return *this;
 }
 
-Vector3D geomath::operator*(const float & lhs, const Vector3D & rhs) {
+Vector3D geomath::operator*(const double & lhs, const Vector3D & rhs) {
 	return { lhs * rhs.x, lhs * rhs.y, lhs * rhs.z };
 }
